Improve malloc_free (#11)
* Collect thread instead of process memory usage.

* Fix format.

* Allow tracing thread with tid.

* Add clang-format.

* Revert "Allow tracing thread with tid."

We can't trace with a tid. Because the memory might be alloced in a
thread but freed in another thread.

This reverts commit 5f2902f6667e9002ab1754c5d849764192488fd8.

* Convert to PID if a TID is assigned.

* Add TID information in malloc_event.

Since the free() might be called in a difference thread from the one
malloc() was called. TID information is added to malloc event so that it
can be used as a key to update the malloc record.

* Move tid_to_pid() to tracelib.